requirements.txt
Django==1.5
South==0.7.6
Werkzeug==0.8.3
dj-database-url==0.2.1
django-braces==1.0.0
django-easyconfig==0.1
django-extensions==1.1.1
feedparser==5.1.3
psycopg2==2.4.6
requests==1.1.0
six==1.3.0
wsgiref==0.1.2
